年翻修改
1、精简nlp模块；
2、azure声音模型列表载入方式修改；
3、增加自动重新载入知识库；
4、提高funasr的稳定性；
5、前端禁止开启后修改

+++ test/funasr/ASR_server.py
@@ -1,14 +1,8 @@
 import asyncio
 import websockets
-import numpy as np
-from queue import Queue
-import threading
 import argparse
 import json
 from funasr import AutoModel
-import wave
-import numpy as np
-import tempfile
 import os
 
 # 设置日志级别
@@ -26,37 +20,60 @@ args = parser.parse_args()
 # 初始化模型
 print("model loading")
 asr_model = AutoModel(model="paraformer-zh", model_revision="v2.0.4",
-                  vad_model="fsmn-vad", vad_model_revision="v2.0.4",
-                  punc_model="ct-punc-c", punc_model_revision="v2.0.4")
+                      vad_model="fsmn-vad", vad_model_revision="v2.0.4",
+                      punc_model="ct-punc-c", punc_model_revision="v2.0.4")
 print("model loaded")
 websocket_users = {}
+task_queue = asyncio.Queue()
 
 async def ws_serve(websocket, path):
     global websocket_users
     user_id = id(websocket)
+    websocket_users[user_id] = websocket
     try:
         async for message in websocket:
             if isinstance(message, str):
                 data = json.loads(message)
                 if 'url' in data:
-                    await process_wav_file(websocket, data['url'], user_id)
+                    await task_queue.put((websocket, data['url']))
+    except websockets.exceptions.ConnectionClosed as e:
+        logger.info(f"Connection closed: {e.reason}")
+    except Exception as e:
+        logger.error(f"Unexpected error: {e}")
     finally:
+        logger.info(f"Cleaning up connection for user {user_id}")
         if user_id in websocket_users:
             del websocket_users[user_id]
+        await websocket.close()
+        logger.info("WebSocket closed")
 
-async def process_wav_file(websocket, message, user_id):
-    wav_path = message
+async def worker():
+    while True:
+        websocket, url = await task_queue.get()
+        if websocket.open:
+            await process_wav_file(websocket, url)
+        else:
+            logger.info("WebSocket connection is already closed when trying to process file")
+        task_queue.task_done()
+
+async def process_wav_file(websocket, url):
+    wav_path = url
     try:
         res = asr_model.generate(input=wav_path)
-        os.remove(wav_path)
-        if 'text' in res[0]: 
-            await websocket.send(res[0]['text'])
+        if res:
+            if 'text' in res[0] and websocket.open:
+                await websocket.send(res[0]['text'])
     except Exception as e:
-        logger.error(f"Error during model.generate: {e}")
+        print(f"Error during model.generate: {e}")
+    finally:
+        if os.path.exists(wav_path):
+            os.remove(wav_path)
 
-def send_recognition_result(websocket, text):
-    asyncio.run_coroutine_threadsafe(websocket.send(text), asyncio.get_event_loop())
+async def main():
+    start_server = websockets.serve(ws_serve, args.host, args.port, subprotocols=["binary"], ping_interval=10)
+    await start_server
+    worker_task = asyncio.create_task(worker())
+    await worker_task
 
-start_server = websockets.serve(ws_serve, args.host, args.port, subprotocols=["binary"], ping_interval=None)
-asyncio.get_event_loop().run_until_complete(start_server)
-asyncio.get_event_loop().run_forever()
+# 使用 asyncio 运行主函数
+asyncio.run(main())
